prediction and UI added

+++ src/textSummarizer/entity/__init__.py
@@ -39,7 +39,7 @@ class ModelTrainerConfig:
     logging_steps: int
     evaluation_strategy: str
     eval_steps: int
-    save_steps: float
+    save_steps: int
     gradient_accumulation_steps: int
